Polish: eliminate duplication, extract shared abstractions, split large files
Extract ScrollState struct from 3 list components (TaskList, ProjectList,
FileExplorer) that had identical cursor/scroll logic. Consolidate duplicate
vt10x terminal rendering into shared vtrender.go with replayVT10X helper.
Merge fgColor/bgColor into parameterized sgrColor. Extract key byte maps
into keybytes.go. Split root.go views into root_views.go (1107→797 lines).
Extract git commands into gitcmd.go. Add borderedPanel and toStringSet
helpers. Use errors.Is for idiomatic error comparison. Extract
determinePostExitStatus pure function and handleConfirmAction shared handler.

Net reduction: -738 lines (981 deleted, 243 added across 23 files).
All tests pass, go vet clean, 3-reviewer independent re-review: APPROVE.
